Change: combine task endpoints
In order to automatic handle tasks by the type of the dataset and to
reduce the amount of redundant code, all task-create-endpoints are
combined into one.
