--- src/api/v1/task/create_table_learn_task.h
@@ -1,44 +1,0 @@
-/**
- * @file        create_table_learn_task.h
- *
- * @author      Tobias Anker <tobias.anker@kitsunemimi.moe>
- *
- * @copyright   Apache License Version 2.0
- *
- *      Copyright 2019 Tobias Anker
- *
- *      Licensed under the Apache License, Version 2.0 (the "License");
- *      you may not use this file except in compliance with the License.
- *      You may obtain a copy of the License at
- *
- *          http://www.apache.org/licenses/LICENSE-2.0
- *
- *      Unless required by applicable law or agreed to in writing, software
- *      distributed under the License is distributed on an "AS IS" BASIS,
- *      WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *      See the License for the specific language governing permissions and
- *      limitations under the License.
- */
-
-#ifndef KYOUKOMIND_CREATE_TABLE_LEARNTASK_H
-#define KYOUKOMIND_CREATE_TABLE_LEARNTASK_H
-
-#include <common.h>
-#include <libKitsunemimiHanamiNetwork/blossom.h>
-
-class Cluster;
-
-class CreateTableLearnTask
-        : public Kitsunemimi::Hanami::Blossom
-{
-public:
-    CreateTableLearnTask();
-
-protected:
-    bool runTask(Kitsunemimi::Hanami::BlossomIO &blossomIO,
-                 const Kitsunemimi::DataMap &context,
-                 Kitsunemimi::Hanami::BlossomStatus &status,
-                 Kitsunemimi::ErrorContainer &error);
-};
-
-#endif // KYOUKOMIND_CREATE_TABLE_LEARNTASK_H
